fix: move keyboard event to components package
