+++ packages/suite/src/views/wallet/transactions/components/TransactionList/components/Actions/components/SearchAction/index.tsx
@@ -10,7 +10,7 @@ import { Account } from '@wallet-types';
 import { Translation } from '@suite-components';
 import { isEnabled } from '@suite-utils/features';
 import { useHotkeys } from 'react-hotkeys-hook';
-import { KEYBOARD_CODE } from '@suite-constants/keyboardEvents';
+import { KEYBOARD_CODE } from '@trezor/components/src/constants/keyboardEvents';
 
 
 const Wrapper = styled.div<{ expanded: boolean }>`
